Ask to destroy DVC if the storage bucket does not match

+++ edge/dvc.py
@@ -1,5 +1,7 @@
 import glob
 import subprocess
+from typing import Optional
+
 import dvc
 import os
 import shutil
@@ -7,20 +9,14 @@ from edge.config import EdgeConfig
 from edge.state import StorageBucketState
 
 
-def dvc_init():
-    print("## Initialising DVC")
-    if os.path.exists(".dvc") and os.path.isdir(".dvc"):
-        print("DVC is already initialised")
-        choice = None
-        while choice not in ["y", "n"]:
-            choice = input("Do you want to destroy DVC and initialise it from scratch (y/n)? [n]: ").strip().lower()
-            if choice == "":
-                choice = "n"
+def dvc_exists() -> bool:
+    return os.path.exists(".dvc") and os.path.isdir(".dvc")
 
-        if choice == "n":
-            return
-        elif choice == "y":
-            dvc_destroy()
+
+def dvc_init():
+    if dvc_exists():
+        return
+    print("## Initialising DVC")
     try:
         subprocess.check_output("dvc init", shell=True)
     except subprocess.CalledProcessError as e:
@@ -37,7 +33,8 @@ def dvc_destroy():
     for f in glob.glob("data/fashion-mnist/*.dvc"):
         os.remove(f)
     print("## Deleting pipeline lock file [models/pipelines/fashion/dvc.lock]")
-    os.remove("models/pipelines/fashion/dvc.lock")
+    if os.path.exists("models/pipelines/fashion/dvc.lock"):
+        os.remove("models/pipelines/fashion/dvc.lock")
     print("DVC is destroyed")
 
 
@@ -55,6 +52,20 @@ def dvc_remote_exists(path: str) -> (bool, bool):
     except subprocess.CalledProcessError as e:
         print(e.output)
         print("Error occurred while adding remote storage to DVC")
+        exit(1)
+
+
+def get_dvc_storage_path() -> Optional[str]:
+    try:
+        remotes_raw = subprocess.check_output("dvc remote list", shell=True).decode("utf-8")
+        remotes = [x.split("\t") for x in remotes_raw.strip().split("\n") if len(x.split("\t")) == 2]
+        for remote in remotes:
+            if remote[0] == "storage":
+                return remote[1].strip()
+        return None
+    except subprocess.CalledProcessError as e:
+        print(e.output)
+        print("Error occurred while getting DVC remote storage path")
         exit(1)
 
 
@@ -77,7 +88,28 @@ def dvc_add_remote(path: str):
 
 
 def setup_dvc(_config: EdgeConfig, storage_state: StorageBucketState):
-    print("# Setting up DVC")
-    dvc_init()
     storage_path = os.path.join(storage_state.bucket_path, _config.storage_bucket.dvc_store_directory)
+    print("# Setting up DVC")
+    print("## Checking if DVC is already initialised")
+    exists = dvc_exists()
+    if exists:
+        print("DVC is initialised")
+        if storage_path != get_dvc_storage_path():
+            print(
+                f"DVC remote storage does not match vertex:edge config: expected -- {storage_path}, "
+                f"got -- {get_dvc_storage_path()}"
+            )
+            print("WARNING: To use the new Google Storage bucket it is advised to destroy DVC repository, "
+                  "and initialise from scratch.")
+            choice = None
+            while choice not in ["y", "n"]:
+                choice = input("Do you want to destroy DVC and initialise it from scratch (y/n)? [n]: ").strip().lower()
+                if choice == "":
+                    choice = "n"
+
+            if choice == "y":
+                dvc_destroy()
+    else:
+        print("DVC is not initialised")
+    dvc_init()
     dvc_add_remote(storage_path)
